factored Param out
